+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 * Initial code release 
 
 ## 0.0.0
+* $this to __CLASS__
 * Add WPS_PLUGIN_DIR path
 * Add `do_load` 
+* Add plugin init load action `wpsp_init`
 * Enable licensing 
 * Make scripts/styles consistent
 * Revise headers